backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
python-dotenv>=1.0.0
paho-mqtt>=2.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
gunicorn>=21.0.0
pytest>=8.0.0
pytest-cov>=4.1.0
